fix: restructure metaplex resolver
